requirements.txt
google-cloud-speech
google-cloud-translate
google-cloud-texttospeech
sounddevice
numpy
python-dotenv
streamlit
streamlit-mic-recorder
pydub